Code organization
Moved part of MCListener
Added support for custom chat player list
